Rename title fields to name
Having the title field on person is non-ideal, title generally refers to
Mr, Mrs, etc.

We don't want to have different names for different types of objects,
makes it hard to display the name of the object.

Using name makes sense everywhere. Now that the filename use of name
has been renamed to filename.